removed hard coding at the service layer

+++ src/main/java/com/nishant/contacts/repository/ContactRepository.java
@@ -8,7 +8,7 @@ import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface ContactRepository extends CrudRepository<Contact, Long> {
+public interface ContactRepository extends CrudRepository<Contact, Integer> {
 
     public List<com.nishant.contacts.entity.Contact> findContactByLastName(String lastName);
